requirements.txt
flask
python-dotenv
flask-wtf
flask-sqlalchemy
flask-migrate
flask-login
aiosmtpd
flask-mail
pyjwt
flask-moment
flask-babel
